requirements.txt
Django==4.1.5
django-allauth==0.52.0
django-cors-headers==4.3.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.4
gunicorn==21.2.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
whitenoise==6.3.0
